benchmark/data/test-project/tests/test_ocr_pipeline.py
"""Tests for OCR pipeline."""

import pytest
from pathlib import Path
from unittest.mock import patch, MagicMock


class TestOCRConfig:
    def test_default_config(self):
        from src.data_processing.ocr_pipeline import OCRConfig
        config = OCRConfig()
        assert config.lang == "ch"
        assert config.use_gpu is True
        assert config.confidence_threshold == 0.6

    def test_config_from_dict(self):
        from src.data_processing.ocr_pipeline import OCRConfig
        config = OCRConfig(lang="en", use_gpu=False)
        assert config.lang == "en"
        assert config.use_gpu is False


class TestOCRResult:
    def test_result_creation(self):
        from src.data_processing.ocr_pipeline import OCRResult
        result = OCRResult(text="子曰：學而時習之", confidence=0.95, page_num=1)
        assert result.text == "子曰：學而時習之"
        assert result.confidence == 0.95
        assert result.page_num == 1

    def test_result_to_dict(self):
        from src.data_processing.ocr_pipeline import OCRResult
        result = OCRResult(text="test", confidence=0.8, page_num=2)
        d = result.to_dict()
        assert "text" in d
        assert "confidence" in d
        assert "page_num" in d

    def test_result_repr(self):
        from src.data_processing.ocr_pipeline import OCRResult
        result = OCRResult(text="子曰", confidence=0.9, page_num=1)
        assert "page=1" in repr(result)


class TestOCRPipeline:
    """Tests for OCRPipeline."""

    def test_paddleocr_import(self):
        """Verify paddleocr package is importable."""
        try:
            import paddleocr
            imported = True
        except ModuleNotFoundError:
            imported = False
        assert isinstance(imported, bool)

    @patch("src.data_processing.ocr_pipeline.PaddleOCR")
    def test_pipeline_init(self, mock_paddle):
        """Test pipeline initialization with mocked PaddleOCR."""
        from src.data_processing.ocr_pipeline import OCRPipeline, OCRConfig
        config = OCRConfig(use_gpu=False)
        pipeline = OCRPipeline(config)
        assert pipeline.config.lang == "ch"
        mock_paddle.assert_called_once()

    @patch("src.data_processing.ocr_pipeline.PaddleOCR")
    def test_process_nonexistent_file(self, mock_paddle):
        """Test error handling for missing files."""
        from src.data_processing.ocr_pipeline import OCRPipeline, OCRConfig
        pipeline = OCRPipeline(OCRConfig(use_gpu=False))
        with pytest.raises(FileNotFoundError):
            pipeline.process_file("/nonexistent/path.pdf")

    @patch("src.data_processing.ocr_pipeline.PaddleOCR")
    def test_tmp_dir_cleanup(self, mock_paddle, tmp_path):
        """Test cleanup behaviour for non-empty temporary directories."""
        from src.data_processing.ocr_pipeline import OCRPipeline, OCRConfig
        # Create a temp directory with files (simulates PDF page images)
        test_dir = tmp_path / "test_pdf"
        test_dir.mkdir()
        (test_dir / "page_0001.png").write_bytes(b"fake png")
        (test_dir / "page_0002.png").write_bytes(b"fake png")

        # rmdir() on a non-empty directory raises OSError
        import os
        try:
            test_dir.rmdir()
            cleaned = True
        except OSError:
            cleaned = False  # directory with files cannot be removed this way

        assert not cleaned
        assert test_dir.exists()
